requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
astroid==2.8.3; python_version ~= '3.6'
certifi==2021.10.8
charset-normalizer==2.0.7; python_version >= '3'
django==2.2
idna==3.3; python_version >= '3'
isort==5.9.3; python_version < '4' and python_full_version >= '3.6.1'
lazy-object-proxy==1.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
mccabe==0.6.1
mysqlclient==2.0.3
numpy==1.21.2
pandas==1.3.4
payjp==0.1.0
pillow==8.4.0
platformdirs==2.4.0; python_version >= '3.6'
pylint-django==2.4.4
pylint-plugin-utils==0.6
pylint==2.11.1; python_version ~= '3.6'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.19.1
python-memcached==1.59
pytz==2021.3
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==3.10.0.2; python_version < '3.10'
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
wrapt==1.13.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
